Added untested IDAStar
Fixed special case for HPAStar

+++ PathFindingTest/Pathfinding.cpp
@@ -82,8 +82,8 @@ int Pathfinding::getNrOfPathNodes() const
 float Pathfinding::getHeuristicDistance(Vec2D start, Vec2D goal) const
 {
 	float h = 0;
-	short x = abs(goal._x - start._x);						//horizontal distance to goal
-	short y = abs(goal._y - start._y);						//vertical distance to goal
+	float x = (float)abs(goal._x - start._x);						//horizontal distance to goal
+	float y = (float)abs(goal._y - start._y);						//vertical distance to goal
 	switch (_heuristicType)
 	{
 	case Pathfinding::MANHATTAN:
